+++ django_athm/signals.py
@@ -6,3 +6,4 @@ athm_response_received = django.dispatch.Signal()
 # Status specific signals
 athm_completed_response = django.dispatch.Signal()
 athm_cancelled_response = django.dispatch.Signal()
+athm_expired_response = django.dispatch.Signal()